massive CLI output updates
