+++ packages/cli/package.json
@@ -73,6 +73,7 @@
     "path-browserify": "^1.0.0",
     "pkg-up": "^3.1.0",
     "prettier": "^1.19.1",
+    "pretty-bytes": "^5.3.0",
     "process": "^0.11.10",
     "regex-parser": "^2.2.10",
     "rollup": "^1.27.10",
