+++ packages/deployer-aws-lambda/src/deploy.ts
@@ -1,76 +1,10 @@
 import { ConfigTypes, FabServerDeployer, FabSettings } from '@fab/core'
 import { createPackage } from './createPackage'
 import path from 'path'
-import fs from 'fs-extra'
 import { _log, FabDeployError } from '@fab/cli'
-import aws, { CloudFront } from 'aws-sdk'
+import { update_lambda, updateCloudFront } from './aws'
 
 const log = _log(`deployer-aws-lambda:deploy`)
-
-const update_lambda = async (
-  package_path: string,
-  accessKeyId: string,
-  secretAccessKey: string,
-  lambda_arn: string,
-  region: string
-) => {
-  const package_contents = await fs.readFile(package_path)
-  const lambda = new aws.Lambda({
-    accessKeyId,
-    secretAccessKey,
-    region,
-  })
-  const params = {
-    FunctionName: lambda_arn,
-    ZipFile: package_contents,
-    Publish: true,
-  }
-  const response = await lambda.updateFunctionCode(params).promise()
-  console.log({ response })
-  return response.Version
-}
-
-const updateCloudFront = async (
-  accessKeyId: string,
-  secretAccessKey: string,
-  lambda_arn: string,
-  cf_distribution_id: string,
-  region: string,
-  version: string
-) => {
-  const cloudfront = new aws.CloudFront({
-    accessKeyId,
-    secretAccessKey,
-    region: 'us-east-1',
-  })
-  const config = await cloudfront
-    .getDistributionConfig({ Id: cf_distribution_id })
-    .promise()
-  console.log({ config })
-  // @ts-ignore
-  console.log(config.DefaultCacheBehavior.LambdaFunctionAssociations)
-  const lambda_config = {
-    Quantity: 1,
-    Items: [
-      {
-        LambdaFunctionARN: `${lambda_arn}:${version}`,
-        EventType: 'origin-request',
-      },
-    ],
-  }
-  // @ts-ignore
-  config.DefaultCacheBehavior.LambdaFunctionAssociations = lambda_config
-  // @ts-ignore
-  console.log(config.DefaultCacheBehavior.LambdaFunctionAssociations)
-  const params = {
-    DistributionConfig: config.DistributionConfig as CloudFront.Types.DistributionConfig,
-    Id: cf_distribution_id,
-    IfMatch: config.ETag,
-  }
-  console.log({ params })
-  const update_response = await cloudfront.updateDistribution(params).promise()
-  console.log({ update_response })
-}
 
 export const deployServer: FabServerDeployer<ConfigTypes.AwsLambda> = async (
   fab_path: string,
