+++ packages/deployer-aws-lambda/src/index.ts
@@ -1,2 +1,4 @@
 export * from './deploy'
 export * from './createPackage'
+export { updateCloudFront } from './aws'
+export { update_lambda } from './aws'
